#20 move test to correct folder
