+++ backend/test/modules/player-profile/core/application/TestPlayerProfileModule.ts
@@ -1,6 +1,6 @@
 import { testModuleCore, TestModuleCore } from '../../../../test-support/shared/core/TestModuleCore';
 import { PlayerProfilesModuleCore } from '../../../../../src/modules/player-profiles/core/PlayerProfilesModuleCore';
-import { InMemoryPlayerProfileRepository } from '../../../../../src/modules/tournaments-registrations/infrastructure/repository/inmemory/InMemoryPlayerProfileRepository';
+import { InMemoryPlayerProfileRepository } from '../../../../../src/modules/player-profiles/infrastructure/repository/inmemory/InMemoryPlayerProfileRepository';
 
 export function testPlayerProfileModule(currentTime: Date): TestModuleCore {
   const playerProfilesRepository = new InMemoryPlayerProfileRepository();
